Revamp EventSelector as a Drawer instead of series of Popovers & Dropdowns.
