+++ frontend/src/kissena/pages/events/EventSelector/Event.selector.module.css
@@ -7,7 +7,3 @@
         border: 2px solid var(--mantine-color-gray-2);
     }
 }
-
-.filter-drawer {
-    /* background-color: var(--mantine-color-darkGreen-0); */
-}
