+++ frontend/src/kissena/pages/events/EventSelector/index.ts
@@ -1,7 +1,6 @@
 import { EventSelector } from "./Event.selector";
-import { EventSelectorMobile } from "./Event.selector.mobile";
 import { EventSelectorContext } from "./Event.selector.context";
 import { EventSelectorProvider } from "./Event.selector.provider";
 import { allTeams } from "./allTeams";
 
-export {EventSelector, EventSelectorMobile, EventSelectorContext, EventSelectorProvider, allTeams };
+export {EventSelector, EventSelector as EventSelectorMobile, EventSelectorContext, EventSelectorProvider, allTeams };
